requirements.txt
keyring
questionary
requests
keyrings.alt
vectra-api-tools